third_party/WebKit/LayoutTests/inspector/console/console-format-es6.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/console-test.js"></script>
<script>

var globals = [];

function log(current)
{
    console.log(globals[current]);
    console.log([globals[current]]);
}

function onload()
{
    var smb1 = Symbol();
    var smb2 = Symbol("a");

    globals = [
        smb1, smb2
    ];

    runTest();
}

function test()
{
    InspectorTest.evaluateInPage("globals.length", loopOverGlobals.bind(this, 0));

    function loopOverGlobals(current, total)
    {
        function advance()
        {
            var next = current + 1;
            if (next == total.description) {
                InspectorTest.expandConsoleMessages();
                InspectorTest.runAfterPendingDispatches(finish);
            } else {
                loopOverGlobals(next, total);
            }
        }

        function finish()
        {
            InspectorTest.dumpConsoleMessages();
            InspectorTest.completeTest();
        }

        InspectorTest.evaluateInPage("log(" + current + ")");
        InspectorTest.runAfterPendingDispatches(evalInConsole);
        function evalInConsole()
        {
            InspectorTest.evaluateInConsole("globals[" + current + "]");
            InspectorTest.runAfterPendingDispatches(advance);
        }
    }
}
</script>
</head>

<body onload="onload()">
<p>
Tests that console properly displays information about ES6 features.
</p>
</body>
</html>
